moved streamlit components onto seperate file
